requirements.txt
Flask>=0.10.1
Jinja2>=2.7.3
Werkzeug>=0.9.6
facebook-sdk>=0.4.0
flask-heroku>=0.1.9
gunicorn>=19.1.1
gevent>=1.0.1
flask_appconfig
